Modified Dockerfile to make java the root process
I modified the Dockerfile to make sure that java is executed as the
root process within the docker container. The reason behind this is
beforehand java was running as a subprocess of the shell which meant
that the java process didn’t receive the kill signal when the docker
container was stopped. This could leave java processes as zombies.

Additionally I introduced a shell script which enables us to provide
additional configuration options via the following optional env variables:

- $JAR_NAME: In case you renamed your jar differently than “microservice-demo-order.jar”
- $EXTRA_JAVA_OPTS: use this variable to overwrite JVM parameters e.g. Heap or GC; leave empty to see get a print of the default values
- $ADDITIONAL_PARAMETERS: Any parameter from application.yaml you want to overwrite
- $PROFILE: The Spring environment profile to be activated. Default: “dev”
- $INSTANCE_NAME: shows up in the process list on linux and is available to your application as a `System.getProperty(“INSTANCE_NAME”)`